model/requirements.txt
# ============================================================
# Real-Time Unusual Activity Detection System
# requirements.txt  (inference + API only — models are pre-trained)
# ============================================================

# --- Core ML Framework ---
torch>=2.1.0
torchvision>=0.16.0

# --- YOLOv8 (includes ByteTrack) ---
ultralytics>=8.1.0

# --- Computer Vision ---
opencv-python>=4.9.0
Pillow>=10.2.0

# --- Numerical ---
numpy>=1.26.0

# --- Configuration ---
PyYAML>=6.0.1

# --- Logging / CLI ---
colorama>=0.4.6
rich>=13.7.0

# --- Optional: ONNX export ---
onnx>=1.15.0
onnxruntime>=1.17.0
# onnxruntime-gpu>=1.17.0  # Uncomment if using CUDA ONNX runtime

# --- Utilities ---
psutil>=5.9.0
python-dotenv>=1.0.0

# --- API Server ---
flask>=3.0.0
flask-cors>=4.0.0
